refactor: rename log_utils to fit the code style
