+++ src/brain/agents/expand_agent.py
@@ -10,7 +10,7 @@ from src.brain.kernel.state_store import (
     save_json_list,
 )
 from src.platform.contracts import CommandSpec
-from src.utils.Logger import get_logger
+from src.utils.log_utils import get_logger
 from src.utils.time_utils import now_text
 
 if TYPE_CHECKING:
