+++ src/brain/agents/plan_agent.py
@@ -9,7 +9,7 @@ from src.brain.kernel.state_store import (
     save_json_list,
 )
 from src.platform.contracts import AppEvent
-from src.utils.Logger import get_logger
+from src.utils.log_utils import get_logger
 from src.utils.time_utils import now_text
 
 if TYPE_CHECKING:
