+++ src/platform/loop.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import asyncio
 
 from src.platform.application_host import ApplicationHost
-from src.utils.Logger import get_logger
+from src.utils.log_utils import get_logger
 
 logger = get_logger("AppLoop")
